// src/config/app-config.type.ts
export type AppConfig = {
  nodeEnv: string;
  name: string;
  workingDirectory: string;
  frontendDomain?: string;
  aeFrontendDomain?: string;
  frontendWebsiteDomain?: string;
  backendDomain: string;
  port: number;
  apiPrefix: string;
  fallbackLanguage: string;
  headerLanguage: string;
  environment: string;
  defaultLiveGroup?: string;
  defaultDemoGroup?: string;
  otpThreshold?: number;
  otpThresholdCount?: number;
  deviceIdOtpThreshold?: number;
  deviceIdOtpThresholdCount?: number;
  crmFrontEndUrl?: string;
  restApiDevUrl?: string;
  restApiQAUrl?: string;
  restApiStagingUrl?: string;
  restApiProdUrl?: string;
  preferredPsp?: string;
  bypassOtpVerification?: string;
  withdrawalMinValue?: number;
  withdrawalMaxValue?: number;
  withdrawalFee?: number;
  depositMinValue?: number;
  depositMaxValue?: number;
  depositFee?: number;
  domain?: string;
  allowedDomains?: string[];
  defaultAggregator?: string;
  defaultCryptoPsp?: string;
  transactionNotPaidDifference?: number;
  financialCCList: string[];
  mt5RegulationServers?: string[];
  transferTransactionType?: string;
  transferTransactionGroupId?: string;
  transferTransactionClientId?: string;
  encryptionKey?: string;
  paymentGatewayCrypto?: string;
  cryptoCurrency?: string;
  cryptoStandard?: string;
  cryptoStandardCond?: string;
  yahooNewsUrl?: string;
  yahooNewsApiKey?: string;
  yahooNewsApiHost?: string;
  companyName?: string;
  btcWalletAddress?: string;
  btcQRCode?: string;
  cryptoMaxDepositValue?: number;
  mt5ManagerLiveUrl?: string;
  mt5ManagerDemoUrl?: string;
  // Migrated from mt5-rest-api by Arshad Shaheen
  symbols?: string[];
  siliconfortMt5ManagerUrl?: string;
  enableDummyData?: boolean;
  bypass2faVerification?: string;
  financialEmailCCExcluseOfficeIds?: number[];
  timezoneDifference?: number;
};